More observable steps (#1)
* Add new substeps for observation purposes

Until now, a MiniJava trace was not providing us with detailed information about some executions steps. For example, in an assignment, it was impossible to know what value was assigned to what variable. The reason was that the specific change occurring in the assignment was capture as a generic "statement step", without any metadata about the endeavors of the statement.

This branch changes the MiniJava semantics and add a wide range of new execution steps (all announted with a comment // Observe only) that make the trace much more interesting. These new steps have parameters that allow us to attach interesting data to them − for instance, parameters telling us what variable was changed in an assignment, and with which new value.

Note that these new steps are in some way "polluting" the step tree, as they are always substeps of previously present steps. But without them it would be required to analyse each pair of states to make such analyses, with the risk of reaching ambiguities.


* Provide model examples in both XMI and Xtext